csv-capture/requirements.txt
hidapi
pycryptodome
